Finished Event Delete tests in controller and Service.

+++ src/main/java/calendar/controller/EventController.java
@@ -59,6 +59,12 @@ public class EventController {
     @RequestMapping(value = "/deleteEvent", method = RequestMethod.DELETE)
     public ResponseEntity<BaseResponse<String>> deleteEvent(@RequestAttribute("userId") int userId, @RequestParam int eventId) {
 
+        User userOfEvent = userService.getById(userId);
+
+        if(userOfEvent == null){
+            return ResponseEntity.badRequest().body(BaseResponse.failure("The user does not exist!"));
+        }
+
         try {
             if (eventService.deleteEvent(eventId) > 0)/* if number of deleted rows in DB > 0 */
                 return ResponseEntity.ok(BaseResponse.success("Event Deleted Successfully"));
